refactor: split positions page and reduce complexity
