+++ packages/frontend/src/pages/positions/useStyles.ts
@@ -1,0 +1,91 @@
+import { createStyles, makeStyles } from '@material-ui/core'
+
+const useStyles = makeStyles((theme) =>
+  createStyles({
+    container: {
+      margin: theme.spacing(6, 8),
+      width: '800px',
+      marginLeft: 'auto',
+      marginRight: 'auto',
+      paddingBottom: theme.spacing(8),
+      [theme.breakpoints.down('sm')]: {
+        width: '100%',
+        padding: theme.spacing(0, 2),
+      },
+    },
+    header: {
+      marginTop: theme.spacing(8),
+      display: 'flex',
+      justifyContent: 'space-between',
+    },
+    position: {
+      padding: theme.spacing(2),
+      backgroundColor: `${theme.palette.background.paper}40`,
+      marginTop: theme.spacing(2),
+      borderRadius: theme.spacing(1),
+      display: 'flex',
+      justifyContent: 'space-between',
+      [theme.breakpoints.down('sm')]: {
+        display: 'block',
+      },
+    },
+    positionData: {
+      display: 'flex',
+      justifyContent: 'space-between',
+      width: '65%',
+      [theme.breakpoints.down('sm')]: {
+        marginTop: theme.spacing(1),
+        width: '100%',
+      },
+    },
+    shortPositionData: {
+      width: '65%',
+      [theme.breakpoints.down('sm')]: {
+        marginTop: theme.spacing(1),
+        width: '100%',
+      },
+    },
+    innerPositionData: {
+      display: 'flex',
+      width: '100%',
+    },
+    positionTitle: {
+      width: '30%',
+      [theme.breakpoints.down('sm')]: {
+        width: '100%',
+      },
+    },
+    empty: {
+      marginTop: theme.spacing(2),
+    },
+    green: {
+      color: theme.palette.success.main,
+    },
+    red: {
+      color: theme.palette.error.main,
+    },
+    history: {
+      marginTop: theme.spacing(8),
+      marginBottom: theme.spacing(8),
+    },
+    link: {
+      color: theme.palette.primary.main,
+      textDecoration: 'underline',
+      fontWeight: 600,
+      fontSize: 14,
+      marginTop: theme.spacing(1),
+    },
+    infoIcon: {
+      fontSize: '10px',
+      marginLeft: theme.spacing(0.5),
+    },
+    tooltipContainer: {
+      marginLeft: '.5em',
+    },
+    dotIcon: {
+      marginRight: '1em',
+    },
+  }),
+)
+
+export default useStyles
